+++ bindings/vue/src/index.js
@@ -47,6 +47,7 @@ const install = (Vue, params = {}) => {
    */
   Vue.prototype.$ons = Object.keys(ons)
     .filter(k => [
+      /^is/,
       /^disable/,
       /^enable/,
       /^open/,