+++ core/src/elements/ons-splitter-side.js
@@ -127,19 +127,36 @@ class CollapseMode {
     }
   }
 
-  _onDragStart(event) {
-    const scrolling = !/left|right/.test(event.gesture.direction);
-    const distance = this._element._side === 'left' ? event.gesture.center.clientX : window.innerWidth - event.gesture.center.clientX;
-    const area = this._element._swipeTargetWidth;
+  _canConsumeGesture(gesture) {
     const isOpen = this.isOpen();
-    this._ignoreDrag = scrolling || (area && distance > area && !isOpen);
+    const validDrag = d => this._element._side === 'left'
+      ? ((d === 'left' && isOpen) || (d === 'right' && !isOpen))
+      : ((d === 'left' && !isOpen) || (d === 'right' && isOpen));
 
-    this._width = widthToPx(this._element._width, this._element.parentNode);
-    this._startDistance = this._distance = isOpen ? this._width : 0;
+    const distance = this._element._side === 'left' ? gesture.center.clientX : window.innerWidth - gesture.center.clientX;
+    const area = this._element._swipeTargetWidth;
+
+    return (validDrag(gesture.direction) || validDrag(gesture.interimDirection)) && !(area && distance > area && !isOpen);
+  }
+
+  _onDragStart(event) {
+    this._ignoreDrag = event.consumed || !this._canConsumeGesture(event.gesture);
+
+    if (!this._ignoreDrag) {
+      event.consume && event.consume();
+      event.consumed = true;
+
+      this._width = widthToPx(this._element._width, this._element.parentNode);
+      this._startDistance = this._distance = this.isOpen() ? this._width : 0;
+
+      util.skipContentScroll(event.gesture);
+    }
   }
 
   _onDrag(event) {
+    event.stopPropagation();
     event.gesture.preventDefault();
+
     const delta = this._element._side === 'left' ? event.gesture.deltaX : -event.gesture.deltaX;
     const distance = Math.max(0, Math.min(this._width, this._startDistance + delta));
     if (distance !== this._distance) {
@@ -150,6 +167,8 @@ class CollapseMode {
   }
 
   _onDragEnd(event) {
+    event.stopPropagation();
+
     const {_distance: distance, _width: width, _element: el} = this;
     const direction = event.gesture.interimDirection;
     const shouldOpen = el._side !== direction && distance > width * el._threshold;
@@ -556,7 +575,7 @@ export default class SplitterSideElement extends BaseElement {
     const action = swipeable === null ? 'off' : 'on';
 
     if (this._gestureDetector) {
-      this._gestureDetector[action]('dragstart dragleft dragright dragend', this._boundHandleGesture);
+      this._gestureDetector[action]('drag dragstart dragend', this._boundHandleGesture);
     }
   }