+++ core/src/ons/notification.js
@@ -84,13 +84,14 @@ notification._createAlertDialog = options => {
   let buttons = '';
   options.buttonLabels.forEach((label, index) => {
     buttons += `
-      <button class="
-        alert-dialog-button
-        ${index === options.primaryButtonIndex ? ' alert-dialog-button--primal' : ''}
-        ${options.buttonLabels.length <= 2 ? ' alert-dialog-button--rowfooter' : ''}
-      ">
+      <ons-alert-dialog-button
+        class="
+          ${index === options.primaryButtonIndex ? ' alert-dialog-button--primal' : ''}
+          ${options.buttonLabels.length <= 2 ? ' alert-dialog-button--rowfooter' : ''}
+        " 
+        style="position: relative;">
         ${label}
-      </button>
+      </ons-alert-dialog-button>
     `;
   });