+++ css-components/gulpfile.js
@@ -10,7 +10,6 @@ const fs = require('fs');
 const ancss = require('ancss');
 const autoprefixer = require('autoprefixer');
 const cssnext = require('postcss-cssnext');
-const immutableCss = require('immutable-css');
 const reporter = require('postcss-reporter');
 
 const prefix = __dirname + '/../build/css/';
@@ -55,7 +54,6 @@ gulp.task('cssnext', ['stylelint'], () => {
     cssnext({
       browsers: ['> 1%', 'last 2 versions', 'Firefox ESR', 'Opera 12.1']
     }),
-    immutableCss(),
     reporter({clearMessage: true, throwError: false})
   ];
   return gulp.src('src/onsen-css-components.css')